Add files via upload
I have designed an Active Notch Filter (ANF) which concentrates on the improving the Q-factor. I have given a noisy heart beat like signal (noisy ECG signal) using the PWL function in spice  with 50Hz added noise, as an input to the ANF and simulated it. The output ECG signal had got ridden of the 50Hz noise (powerline/mains noise) and appeared to be clean. The strong selective frequency attenuating nature of the designed ANF suppressed the 50Hz frequency component (noise) very strongly as shown in one of the above files.