Add files via upload
A 5T-OTA has been designed using 180 nm technology, featuring an NMOS differential pair with an active PMOS load. The OTA provides a gain of 70 (37 dB), a 3‑dB bandwidth of approximately 9.65 kHz, and a unity-gain bandwidth of around 685 kHz. It includes a tail current source delivering a constant 1 µA, operates with a 1.4 V power supply, and is biased at a common-mode voltage of 1.2 V.